Title: 산타의 선물 공장, Time: 182ms, Memory: 8MB, Status: Passed - CodeTree

+++ 240408/산타의 선물 공장/santa-gift-factory.cpp
@@ -67,7 +67,7 @@ int main() {
         }
         if(op==200){
             int w_max; cin>>w_max;
-            int res=0;
+            long long res=0;
             for (int i = 0; i < m; ++i) {
                 if(belt[i].broken) continue;
                 int head=belt[i].head;
